requirements.txt
streamlit
torch
torchvision
Pillow
joblib
numpy
opencv-python-headless
streamlit-webrtc
scikit-learn